Add a new configuration option: sentryDSN
Adding a new config option, ``sentryDSN`` to provide the Sentry DSN if
the user wants to monitor the BGPalerter app for runtime exceptions.

I have updated the package.json file to the version recommended in
https://docs.sentry.io/error-reporting/quickstart/?platform=node.

This is an optional config bit that enables Sentry monitoring only when
``sentryDSN`` is provided.